app/requirements.txt
Django>=5.2
gunicorn
dnspython>=2.4
